{{cookiecutter.service_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.service_name}}"
version = "0.1.0"
description = "{{cookiecutter.service_description}}"
authors = ["{{cookiecutter.service_authors}}"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.8.2"
databases = {extras = ["postgresql"], version = "^0.4.3"}
requests = "^2.26.0"
fastapi = "^0.67.0"
PyYAML = "^5.4.1"
types-PyYAML = "^5.4.3"
dependency-injector = "^4.35.2"
alembic = "^1.6.5"
psycopg2 = "^2.9.1"
orjson = "^3.6.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
coverage = {extras = ["toml"], version = "^5.5"}
pytest-cov = "^2.12.1"
black = "^21.7b0"
flake8 = "^3.9.2"
flake8-bandit = "^2.1.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.4.3"
flake8-import-order = "^0.18.1"
safety = "^1.10.3"
pre-commit = "^2.13.0"
mypy = "^0.910"
darglint = "^1.8.0"
flake8-docstrings = "^1.6.0"
pytest-mock = "^3.6.1"
poetry2setup = "^1.0.0"
sqlalchemy-stubs = "^0.4"

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["{{cookiecutter.service_name}}"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
